Nba cup + rename Siximpl to SixImpl
